% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{linear_model}
\alias{linear_model}
\title{Fit a linear model}
\usage{
linear_model(response, covariates, alpha = 0.05)
}
\arguments{
\item{response}{A \code{vector} representing the response variable.}

\item{covariates}{A \code{matrix} representing the covariates.}

\item{alpha}{The \code{numeric} value representing the significance level.}
}
\value{
A \code{list} containing :
\describe{
    \item{beta}{Estimated coefficient, beta hat}
    \item{sigma2}{Residual variance (MSE)}
    \item{ci}{confidence interval, derived from alpha}
    \item{resid}{residuals}
    \item{fitted}{fitted values}
    \item{r2}{R squared}
    \item{cp}{Mallows' Cp}
    \item{f}{F statistic}
    \item{pval}{P-Value}
}
}
\description{
Simple function that computes a linear regression model.
Reference https://smac-group.github.io/ds/section-functions.html#section-example-continued-least-squares-function
}
\examples{
linear_model(cars$speed, cars$distance)
linear_model(cars$speed, cars$distance, 0.025)
}
\references{
SMAC Group Book: https://smac-group.github.io/ds/section-functions.html#section-example-continued-least-squares-function.
}
\author{
Group 09(Alex, Nan, Khanh - Stat 6210 Auburn Univ)
}
