% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{model_plots}
\alias{model_plots}
\title{Plot Model diagnostics}
\usage{
model_plots(model)
}
\arguments{
\item{model}{A \code{list} from \code{linear_model} output.}
}
\description{
Function that takes in linear model attributes in form of list and produces several plots:

Residual Vs Fitted Values

QQPlot of Residuals

Histogram of Residuals
}
\examples{
model_plots(model_out)
}
\author{
Alex, Nan, and Khanh - Stat 6210 Auburn Univ
}
