% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{linear_model}
\alias{linear_model}
\title{LinearRegressionNNY}
\usage{
linear_model(response, covariates, alpha = 0.05)
}
\arguments{
\item{response}{A \code{vector} representing the response variable.}

\item{covariates}{A \code{matrix} representing the covariates.}

\item{alpha}{The \code{numeric} value representing the significance level.}
}
\value{
A \code{list} containing :
\describe{
    \item{beta}{Estimated coefficient, beta hat}
    \item{sigma2}{Residuals}
    \item{variance_beta}{variance of the beta estimator, beta hat}
    \item{ci}{confidence interval, derived from alpha}
}
}
\description{
Simple function that computes a linear regression model.
Citation https://smac-group.github.io/ds/section-functions.html#section-example-continued-least-squares-function
}
\examples{
linear_model(cars$speed, cars$distance)
linear_model(cars$speed, cars$distance, 0.025)
}
\author{
Group 09 and SMAC Group: https://smac-group.github.io/ds/section-functions.html#section-example-continued-least-squares-function.
}
